+++ docs/source/base/release_notes.md
@@ -199,7 +199,7 @@ Version 5.2.0: Data Store and Various Improvements
   Commit: 83fdd0c128a3488b7845d474bd609bd004fa7adb
 - Update to the changelog table
   - Make 'author' and 'severity' properties orderable (3436cc545594b4e9379219cf5651170a8acff892)
-  - Add new 'Reversible' boolean property (af0886d0df5f1c807b353d4f191d8ee9faeddeac)
+  - Add new 'Revertible' boolean property (af0886d0df5f1c807b353d4f191d8ee9faeddeac)
 - Display the network/workflow tree by default if it was previously activated.
   The setting is stored in the database and used to automatically show the tree in the HTML template.
   Commit: 0513847d23e9bfd31fbfd22cdfe9137faf5de6a0